+++ app/jobs/type/route.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import Job from '../../models/job';
+import Job from '../../models/job-non-model';
 
 export default Ember.Route.extend({
